Refactoring : New Joint class
Handles all that's joint dependent only with no dependencies on the other joints
